JBRULES-3223 Enable marshalling of sliding time and length windows (cherry picked from commit b869611e377e9fc5e036c64c296eeaba75a5cd0e)

+++ drools-core/src/main/java/org/drools/common/RuleFlowGroupImpl.java
@@ -292,7 +292,7 @@ public class RuleFlowGroupImpl
         }
 
         public void write(MarshallerWriteContext context) throws IOException {
-            context.writeInt( WorkingMemoryAction.DeactivateCallback );
+            context.writeShort( WorkingMemoryAction.DeactivateCallback );
             context.writeUTF( ruleFlowGroup.getName() );
         }
